requirements.txt
streamlit==1.35.0
streamlit-drawable-canvas
Pillow
numpy